refactor: split inline helpers
